Support old versions of Windows that doesn't have built-in bsdtar
Built-in bsdtar is only available on Windows 10 19H1
(Version 1903, Build 18362) or higher.[1]
[1] Insider Build: Build 17063 or higher

Sadly, a lot of people even using Windows 10 don't update to
the latest version and use an older version of Windows.
In this case, they can install 7-Zip and use it to extract
vendor.img file manually. But this can be a hassle.

So for those who use an old version of Windows that
doesn't have built-in bsdtar, this commit adds a feature
to automatically download 7za.exe (version 19.00, x86 binary)
and use it to decompress tgz files.

Of course, when running on Windows 10 Build 17063 or higher,
7za.exe is not download and the built-in bsdtar is still used.