Use small byte buffer for extracting payload.tgz
Previously, the entire sh file was loaded into memory,
and the most of the sh file (except lines 1 to 314)
were saved as payload.tgz.
Since the sh file is over 200 MiB, this will use up
too much memory unnecessarily. So this is not a good idea.

A much smaller byte buffer is enough to work.

I checked various size of the byte buffer, and found that
64 KiB (=65536 bytes) is optimal.